% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_manifest}
\alias{load_manifest}
\title{Load manifest information for Illumina Infinium Global Screening Array v3.0 SNP beadchip}
\format{
A data frame with 654,027 rows (probes) and 22 variables:
\describe{
\item{Name}{Name of probe}
\item{IlmnID}{Illumina ID of probe}
\item{RsID}{rs number of probe}
\item{Chr}{Chromosome}
\item{Pos_hg38}{Position on GRCh38}
\item{Pos_hg19}{Position on GRCh37/hg19}
\item{AddressA_ID}{Address A of probe}
\item{AddressB_ID}{Address B of probe}
\item{GenomeBuild}{Genome build, e.g. 38}
\item{SNP}{The possible bases of the SNP}
\item{IlmnStrand}{\code{TOP}/\code{BOT}/\code{PLUS}/\code{MINUS}}
\item{SourceStrand}{\code{TOP}/\code{BOT}/\code{PLUS}/\code{MINUS}}
\item{TopGenomicSeqSBE}{The content of the \verb{.[.].} part of \code{TopGenomicSeq} (not included), including the flanking left and right parts}
\item{TopGenomicSeqSBE_Left}{The left flanking of \code{TopGenomicSeqSBE}}
\item{TopGenomicSeqSBE_Right}{The right flanking of \code{TopGenomicSeqSBE}}
\item{BeadSetID}{Bead set ID}
\item{Exp_Clusters}{3 or 2}
\item{RefStrand}{\code{+}/\code{-}}
\item{ProbeType}{\code{II} (only address A) or \code{I} (address A and B)}
\item{SNPType}{Signals whether \code{SNP} has alleles on same color channel (ambiguous) or not (unambiguous), or potentially is an INDEL (insertion/deletion): \code{AMB}iguous for AT-SNPs (\verb{[A/T]} or \verb{[T/A]}) and CG-SNPs; \code{INDEL} for INDELs; \code{UNAMB}iguous for the rest (e.g. AC-SNPs)}
\item{Exclude}{Some probes should be excluded (cf. above), this indicate those}
\item{MappingComment}{Mapping comment - ambiguous mapping, PAR regions etc.}
}
}
\source{
\url{https://sapac.support.illumina.com/array/array_kits/infinium-global-screening-array/downloads.html>}
}
\usage{
load_manifest()
}
\description{
This dataset is a combination of the following four datasets
from https://sapac.support.illumina.com/array/array_kits/infinium-global-screening-array/downloads.html:
}
\details{
\enumerate{
\item \verb{GSA-24v3-0_A2.csv}
\itemize{
\item \verb{Infinium Global Screening Array v3.0 Product Files -> Infinium Global Screening Array v3.0 Manifest File (CSV Format - GRCh38)}
}
\item \verb{GSA-24v3-0_A1.csv}
\itemize{
\item \verb{Infinium Global Screening Array v3.0 Product Files -> Infinium Global Screening Array v3.0 Manifest File (CSV Format - GRCh37)}
}
\item \verb{GSA-24v3-0_A1_b151_rsids.txt}
\itemize{
\item \verb{Infinium Global Screening Array v3.0 Support Files -> Infinium Global Screening Array v3.0 Loci Name to rsID Conversion File}
}
\item \verb{GSA-24v3-0_A1_MappingComment.txt}
\itemize{
\item \verb{Infinium Global Screening Array v3.0 Support Files -> Infinium Global Screening Array v3.0 Mapping Comments}
}
}
}
